# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Translation engine to use (openai, google, pollinations)
 */
export type ProcessVideoRequestTranslationEngine = typeof ProcessVideoRequestTranslationEngine[keyof typeof ProcessVideoRequestTranslationEngine];


export const ProcessVideoRequestTranslationEngine = {
  openai: 'openai',
  google: 'google',
  pollinations: 'pollinations',
} as const;

export interface ProcessVideoRequest {
  /** YouTube video URL */
  videoUrl: string;
  /** Start time in seconds */
  startTime: number;
  /** Edge TTS voice name */
  voice: string;
  /** Speech speed multiplier (auto-calculated if omitted) */
  speed?: number;
  /** Translation engine to use (openai, google, pollinations) */
  translationEngine?: ProcessVideoRequestTranslationEngine;
  /** Skip captions and extract audio directly for AI transcription */
  forceAudioExtraction?: boolean;
}

export interface ProcessVideoResponse {
  jobId: string;
  status: string;
  message: string;
}

export type JobStatusResponseStatus = typeof JobStatusResponseStatus[keyof typeof JobStatusResponseStatus];


export const JobStatusResponseStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface JobStatusResponse {
  jobId: string;
  status: JobStatusResponseStatus;
  progress: string;
  audioUrl?: string | null;
  transcript?: string | null;
  translation?: string | null;
  error?: string | null;
  startTime?: number | null;
  suggestedRate?: number | null;
}

export interface TtsVoice {
  id: string;
  name: string;
  lang: string;
}

export interface TtsModelsResponse {
  voices: TtsVoice[];
}

export interface CookiesStatusResponse {
  hasCookies: boolean;
}

export interface SaveCookiesRequest {
  cookies: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}