# lib/api-zod/src/generated/types/jobStatusResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusResponseStatus } from './jobStatusResponseStatus';

export interface JobStatusResponse {
  jobId: string;
  status: JobStatusResponseStatus;
  progress: string;
  audioUrl?: string | null;
  transcript?: string | null;
  translation?: string | null;
  error?: string | null;
  startTime?: number | null;
  suggestedRate?: number | null;
}