# lib/api-zod/src/generated/types/processVideoRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessVideoRequestTranslationEngine } from './processVideoRequestTranslationEngine';

export interface ProcessVideoRequest {
  /** YouTube video URL */
  videoUrl: string;
  /** Start time in seconds */
  startTime: number;
  /** Edge TTS voice name */
  voice: string;
  /** Speech speed multiplier (auto-calculated if omitted) */
  speed?: number;
  /** Translation engine to use (openai, google, pollinations) */
  translationEngine?: ProcessVideoRequestTranslationEngine;
  /** Skip captions and extract audio directly for AI transcription */
  forceAudioExtraction?: boolean;
}