# lib/api-zod/src/generated/types/ttsModelsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TtsVoice } from './ttsVoice';

export interface TtsModelsResponse {
  voices: TtsVoice[];
}