# lib/api-zod/src/generated/types/ttsVoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TtsVoice {
  id: string;
  name: string;
  lang: string;
}